### 2019-10-31 #1846 @gkthiruvathukal
@Kevin-Mattheus-Moerman Would be great if you can take this one. I'll be happy to take new ones starting next week.  Going through a bit of a rough stretch with the continued teachers' strike in Chicago and trying to keep middle/high school kids entertained while staying on top of my own work. So I should be in the clear by next week (I hope).
