### 2019-10-31 #1846 @Kevin-Mattheus-Moerman
@hooman650 would you be interested in reviewing this work for JOSS?
