### 2019-11-01 #1846 @Kevin-Mattheus-Moerman
@paulvangentcom @J535D165 @DominiqueMakowski @ismorphism @PGomes92 would you be interested in reviewing this submission for [JOSS](https://joss.theoj.org/)? It concerns a Python package for Heart Rate Variability Analysis.

This is a pre-review issue which we use to recruit reviewers. Once we have sufficient reviewers we open a dedicated review issue. The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) focuses on [the software ](https://github.com/rhenanbartels/hrv) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01846/joss.01846/10.21105.joss.01846.pdf).
Here is an example of a review process where  you can see that checkboxes guide the reviewers through the review process: https://github.com/openjournals/joss-reviews/issues/1745

Let me know if you have questions.
