### 2019-11-01 6d1921d
Return fig and ax objects in hist method
