+++ tests/test_rri.py
@@ -454,3 +454,11 @@ class TestRRiPlotMethods:
         rri = RRi(FAKE_RRI, time=[4, 5, 6, 7])
 
         rri.plot(ax=ax)
+
+    def test_return_fig_and_ax_objects_with_hist(self):
+        rri = RRi(FAKE_RRI, time=[4, 5, 6, 7])
+
+        fig, ax = rri.hist()
+
+        assert isinstance(fig, matplotlib.figure.Figure)
+        assert isinstance(ax, matplotlib.figure.Axes)
