### 2019-11-01 1f2b79e
Mock matplotlib show method
