### 2019-11-01 0daf7b9
Improve ax reuse test
