### 2019-11-01 0df6075
Move test rri to set_up_class
