### 2019-11-01 a1b51a7
Start poincaré plot
