### 2019-11-01 8a32d12
Set xlabel, ylabel and title on poincaré plot
