### 2019-11-06 #1867 @whedon
[REVIEW]: HRV: a Pythonic package for Heart Rate Variability Analysis
**Submitting author:** @rhenanbartels (<a href="http://orcid.org/0000-0002-6088-2186">Rhenan Bartels</a>)
**Repository:** <a href="https://github.com/rhenanbartels/hrv" target ="_blank">https://github.com/rhenanbartels/hrv</a>
**Version:** 0.2.8
