### 2019-11-06 #1867 @Kevin-Mattheus-Moerman
@hooman650, @PGomes92, @paulvangentcom thanks for agreeing to review this work. You have instructions and check-boxes at the top of this issue to guide you through the process.
You can comment on this submission in this review issue. For longer/larger issue we suggest you open an issue on the project repository instead and link to it here. We also encourage you to make use of checkboxes in your comments so it is easy to keep track of steps/comments the author still needs to work on. You can do this using `- [ ] `.

Let me know if you have questions.
