+++ hrv/utils.py
@@ -93,3 +93,17 @@ def _interp_linear(rri, time, fs):
 def _create_interp_time(time, fs):
     time_resolution = 1 / float(fs)
     return np.arange(0, time[-1] + time_resolution, time_resolution)
+
+
+def _ellipsedraw(ax, a, b, x0, y0, phi, *args, **kwargs):
+    theta = np.arange(-0.03, 2*np.pi, 0.01)
+    x = a * np.cos(theta)
+    y = b * np.sin(theta)
+    X = np.cos(phi) * x - np.sin(phi) * y
+    Y = np.sin(phi) * x + np.cos(phi) * y
+
+    X = X + x0
+    Y = Y + y0
+
+    ax.plot(X, Y, *args, **kwargs)
+    return ax
