### 2019-11-07 06357ef
Return fig and ax objects
