+++ hrv/rri.py
@@ -160,6 +160,8 @@ class RRi:
             linewidth=3
         )
 
+        return fig, ax
+
     # TODO: Create methods for time domain to be calculted in the instance
 
     def mean(self, *args, **kwargs):
