### 2019-11-17 rhenanbartels/hrv#16 @hooman650
Artifact removal, outlier detection and baseline wander removal 
Hi,

Thanks for the hrv toolbox. I have noticed that there are numerous HRV toolboxes that perform the same function (computing time, frequency and nonlinearity parameters). But what I find missing in all of these are automated methods to remove outliers in RR intervals and also the baseline wandering. These substantially change the computed outcomes. I suggest looking into the page 11 of the document below for some ideas in this regard:
https://www.kubios.com/downloads/Kubios_HRV_Users_Guide.pdf

Also look into the paper below, how smoothness priors could be used to remove baseline wandering:
https://pdfs.semanticscholar.org/7b28/41f2b7013eccb3b75f72a1aada97ab66d551.pdf

These are relatively simple to implement but having these would make this toolbox definitely much more useful than the other very many ones that do the same operations.

Thanks
Hooman
