### 2019-11-17 #1867 @hooman650
I have opened an issue in the toolbox repository. There are many HRV toolboxes available that exactly perform the same computations (time, frequency and nonlinearity analysis).  I recommend adding a few subroutines for preprocessing RR intervals and removing artifacts. I have opened an [issue](https://github.com/rhenanbartels/hrv/issues/16#issue-524008155) and made a few recommendations there.

Additionally, contribution guidelines are missing.
