### 2019-11-19 rhenanbartels/hrv#16 @rhenanbartels
Hi @hooman650, thank you very much for the suggestions.
I will work on them during the weekend.

Cheers!
