### 2019-11-29 b4c3432
Show poincare legend as SD1 and SD2
