### 2019-11-29 ffbaf72
Add contributing section to README
