### 2019-12-02 1b628b9
Start threshold filter
