+++ tests/test_filters.py
@@ -2,7 +2,11 @@ from unittest import TestCase
 
 import numpy as np
 
-from hrv.filters import moving_average, moving_median, quotient
+from hrv.filters import (
+    moving_average,
+    moving_median,
+    quotient,
+    threshsold_filter)
 from hrv.rri import RRi
 
 
@@ -85,6 +89,24 @@ class Filter(TestCase):
         rri_filt = moving_median(fake_rri)
 
         expected_rri = [810, 830, 830, 804, 801, 801, 800]
+        expected_time = [0, 1, 2, 3, 4, 5, 6]
+
+        assert isinstance(rri_filt, RRi)
+        np.testing.assert_almost_equal(
+                rri_filt.values,
+                expected_rri,
+                decimal=2
+        )
+        np.testing.assert_almost_equal(rri_filt.time, expected_time, decimal=2)
+
+    def test_threshold_filter(self):
+        fake_rri = RRi(
+                [810, 830, 860, 865, 804, 1100, 800],
+                time=[0, 1, 2, 3, 4, 5, 6]
+        )
+
+        rri_filt = threshsold_filter(fake_rri, threshold=250)
+        expected_rri = [810, 830, 860, 865, 804, 748.40625, 800]
         expected_time = [0, 1, 2, 3, 4, 5, 6]
 
         assert isinstance(rri_filt, RRi)
