### 2019-12-02 7b38be9
Filter the beginning part of rri series
