+++ tests/test_filters.py
@@ -116,3 +116,22 @@ class Filter(TestCase):
                 decimal=2
         )
         np.testing.assert_almost_equal(rri_filt.time, expected_time, decimal=2)
+
+    def test_threshold_filter_noise_in_the_beginning(self):
+        fake_rri = RRi(
+                [810, 500, 860, 865, 804, 810, 800],
+                time=[0, 1, 2, 3, 4, 5, 6]
+        )
+
+        rri_filt = threshsold_filter(fake_rri, threshold=250)
+        expected_rri = [810, 814.34375, 860, 865, 804, 810, 800]
+        expected_time = [0, 1, 2, 3, 4, 5, 6]
+
+        assert isinstance(rri_filt, RRi)
+        np.testing.assert_almost_equal(
+                rri_filt.values,
+                expected_rri,
+                decimal=2
+        )
+        np.testing.assert_almost_equal(rri_filt.time, expected_time, decimal=2)
+
