### 2019-12-02 347192b
Add filter strength with string names
