+++ tests/test_filters.py
@@ -135,3 +135,21 @@ class Filter(TestCase):
         )
         np.testing.assert_almost_equal(rri_filt.time, expected_time, decimal=2)
 
+    def test_threshold_filter_string_threshold(self):
+        fake_rri = RRi(
+                [810, 650, 860, 865, 804, 810, 800],
+                time=[0, 1, 2, 3, 4, 5, 6]
+        )
+
+        rri_filt = threshsold_filter(fake_rri, threshold='strong')
+        expected_rri = [810, 814.34375, 860, 865, 804, 810, 800]
+        expected_time = [0, 1, 2, 3, 4, 5, 6]
+
+        assert isinstance(rri_filt, RRi)
+        np.testing.assert_almost_equal(
+                rri_filt.values,
+                expected_rri,
+                decimal=2
+        )
+        np.testing.assert_almost_equal(rri_filt.time, expected_time, decimal=2)
+
