### 2019-12-02 80e2d44
Correct typo in function name
