### 2019-12-03 b809221
Set string medium as default for threshold filter
