### 2019-12-03 ac0eede
Add threshold filter to README
