### 2019-12-03 71cc9e9
Create first polynomial detrend
