### 2019-12-03 666c2da
Start detrended RRi class
