+++ tests/test_rri.py
@@ -6,6 +6,7 @@ import numpy as np
 import pytest
 
 from hrv.rri import (RRi,
+                     RRiDetrended,
                      _validate_rri,
                      _create_time_array,
                      _validate_time,
@@ -515,3 +516,14 @@ class TestRRiPlotMethods:
 
         assert returned_fig == fig_mock
         assert returned_ax == ax_mock
+
+
+class TestRRiDetrended:
+    def test_create_detrended_rri_class(self):
+        detrended_rri = [-87.98470401,  -88.22253018,  -49.46831978,
+                         -109.69798458, -181.90892056]
+        rri_time = [0, 1, 2, 3, 4]
+
+        det_rri_obj = RRiDetrended(detrended_rri, time=rri_time)
+
+        assert det_rri_obj.detrended == True
