### 2019-12-04 23c5118
Remove hardcoded interp and detrended from RRi base class
