### 2019-12-04 b35ae1c
Return RRiDetrended instance and assert average close to zero
