### 2019-12-04 dad76c5
Add forgotten TestCase method
