### 2019-12-04 c9f11b9
Detrend others rri values as normal iterables
