### 2019-12-04 a298d0b
Calculate smoothness priors detrend
