### 2019-12-04 0a4ea66
Implement savtizky_golay detrend
