### 2019-12-04 d96bf45
Correct smoothness priors test
