### 2019-12-04 b18b181
Add default polynomial degree in detrend
