### 2019-12-04 1567b68
Do not detrend in Welch's method if already detrended
