### 2019-12-04 4a20873
Do not interpolated rri if already interpolated
