### 2019-12-04 4a809c2
Detrend with polynomial method when AR method is selected
