### 2019-12-04 05b5b80
Minor code refactor
