### 2019-12-05 29e826b
Add new preprocessing techniques to README
