### 2019-12-05 14bbf36
Add rri detrended figures to README
