### 2019-12-05 2973643
Add a note of matrix inversion issue
