### 2019-12-05 939d49b
Add references to README
