### 2019-12-05 6a7d683
Minor corrections in README
