### 2019-12-05 rhenanbartels/hrv#16 @rhenanbartels
Hi @hooman650, following your suggestions, I've implemented the following features regarding artifact removal and baseline trend removal:

#### Artifact removal
-  ```threshold_filter```: inspired in Kubios **threshold based artefact correction algorithm**, each RRi is compared to a local value consisting of the median of adjacent RRi. If the difference between RRi and the local median is greater than a given threshold in milliseconds this RRi is considered an ectopic beat. Ectopic RRi values are replaced with cubic spline interpolation of the entire tachogram.

It worth noticing that the previous version also offered 3 functions to remove artifacts:
 - moving average filter
 - moving median fitler
 - Quotient filter

#### Detrending functions
In order to remove simple slow or more complex trends, I've implemented 3 functions to remove low-frequency components from the RRi series:

  - ```polynomial_detrend```: fits a polynomial model of a given degree to the tachogram and remove it from each RRi values.
  - ```smoothness_priors```: widely used in hrv analyses, acts as a lowpass filter and remove complex trends from the RRi series
 - ```sg_detrend```: uses the savitzky-golay lowpass filter to remove low-frequency trends from the RRi series

With the detrend methods, I have also created an ```RRiDetrend``` class, which has the same behavior of base ```RRi``` class, but skip some validations of the RRi values, such as the tachogram can only have positive values.

Best regards,

Rhenan
