### 2019-12-09 #1867 @paulvangentcom
I think the paper can benefit from additional integration with existing literature, especially a more recent state of the field.

Although many toolboxes exist to work with PPG, ECG, or the resulting tachograms, I think _hrv_ can play a role here as an accessible toolbox for those already having tachograms available and only need to analyse these.

The software needs to be documented with proper docstrings in the code, and the documentation be migrated to for example readthedocs. Unit tests are missing as is continuous integration and an overview of the Python versions that are supported.

I'll open issues for this on the repo.
