### 2019-12-09 rhenanbartels/hrv#17 @paulvangentcom
Review suggestions
Hi @rhenanbartels,

I was asked to review your submission to JOSS. Some suggestions from my end:

- I think the paper can benefit from additional integration with existing literature, especially a more recent state of the field.

- The software needs to be documented with proper docstrings in the code, I personally recommend either the Numpy or Google style. [see for example docstrings here](https://github.com/numpy/numpy/blob/master/numpy/polynomial/hermite.py)

-  Unit tests are missing, when using Numpy or Google style you can include examples in the docstring that double as unit tests.

- Continuous integration, code coverage metrics, and an overview of the Python versions that are supported is missing too. There's many services out there for unit tests, I personally use travis-ci for CI and codecov.io for coverage results. You can add badges to the repo that update with CI and coverage results for each git push.

Although many toolboxes exist to work with PPG, ECG, or the resulting tachograms, I think _hrv_ can play a nice role as an accessible toolbox for those already having tachograms available and need to analyse these. Congratulations on the work. Please keep me updated on the progress.

\-Paul
