### 2019-12-09 rhenanbartels/hrv#17 @paulvangentcom
One final addition: I suggest adding some demo data to the package and docs, so that everyone can follow along with the quickstart.
