### 2019-12-09 rhenanbartels/hrv#17 @rhenanbartels
Hi @paulvangentcom, thank you very much for the suggestions.

I couldn't agree more. Now it's time to focus on the documentation of the package and CI.

I will work on this issue along the week and send an update when it's done or if I have any doubts.

Best regards,

Rhenan
