### 2019-12-09 27078c1
Add docstring from time_domain method
