### 2019-12-10 a175dcf
Start frequency domain docstring
