### 2019-12-10 c19a50b
Improve frequency-domain function documentation
