### 2019-12-10 733ec39
Add docstring for the non_linear function
