### 2019-12-11 37e5447
Add docstring for hrv filters
