+++ hrv/filters.py
@@ -5,10 +5,41 @@ from hrv.rri import RRi
 from hrv.utils import _create_time_info
 
 
+__all__ = ['quotient', 'moving_average', 'moving_median', 'threshold_filter']
+
+
 def quotient(rri):
+    """
+    Remove ectopic beats which the variation of two consecutive values
+    exceeds 20%
+
+    Parameters
+    ----------
+    rri : array_like
+        sequence containing the RRi series
+
+    Returns
+    -------
+    results : RRi array
+        instance of the RRi class containing the filtered RRi values
+
+    References
+    ----------
+    - Piskorski J, Guzik P. Filtering poincare plots.
+      Comput Methods Sci Technol. 2005;11(1):39–48.
+
+    Examples
+    --------
+    >>> from hrv.filters import quotient
+    >>> from hrv.io import read_from_text
+    >>> rri = read_from_text('/path/to/file.txt')
+    >>> quotient(rri)
+    RRi array([1114., 1113., 1066., 1119., 1062.,])
+    """
     # TODO: Receive option to replaced outliers with stats
     # functions (i.e mean, median etc)
     # TODO: Receive option to re-create time array with cumsum of filtered rri
+    # TODO: Receive threshold value
 
     if isinstance(rri, RRi):
         rri_time = rri.time
@@ -29,14 +60,130 @@ def quotient(rri):
 
 
 def moving_average(rri, order=3):
+    """
+    Low-pass filter. Replace each RRi value by the average of its ⌊N/2⌋
+    neighbors. The first and the last ⌊N/2⌋ RRi values are not filtered
+
+    Parameters
+    ----------
+    rri : array_like
+        sequence containing the RRi series
+    order : int, optional
+        Strength of the filter. Number of adjacent RRi values used to calculate
+        the average value to replace the current RRi. Defaults to 3.
+
+    .. math::
+        considering movinge average of order equal to 3:
+            RRi[j] = sum(RRi[j-2] + RRi[j-1] + RRi[j+1] + RRi[j+2]) / 3
+
+    Returns
+    -------
+    results : RRi array
+        instance of the RRi class containing the filtered RRi values
+
+    See Also
+    -------
+    moving_median, threshold_filter, quotient
+
+    Examples
+    --------
+    >>> from hrv.filters import moving_average
+    >>> from hrv.io import read_from_text
+    >>> rri = read_from_text('/path/to/file.txt')
+    >>> moving_average(rri)
+    RRi array([1114.        , 1097.66666667, 1099.33333333, 1082.33333333,])
+    """
     return _moving_function(rri, order, np.mean)
 
 
 def moving_median(rri, order=3):
+    """
+    Low-pass filter. Replace each RRi value by the median of its ⌊N/2⌋
+    neighbors. The first and the last ⌊N/2⌋ RRi values are not filtered
+
+    Parameters
+    ----------
+    rri : array_like
+        sequence containing the RRi series
+    order : int, optional
+        Strength of the filter. Number of adjacent RRi values used to calculate
+        the median value to replace the current RRi. Defaults to 3.
+
+    .. math::
+        considering movinge average of order equal to 3:
+            RRi[j] = np.median([RRi[j-2], RRi[j-1], RRi[j+1], RRi[j+2]])
+
+    Returns
+    -------
+    results : RRi array
+        instance of the RRi class containing the filtered RRi values
+
+    See Also
+    -------
+    moving_average, threshold_filter, quotient
+
+    Examples
+    --------
+    >>> from hrv.filters import moving_median
+    >>> from hrv.io import read_from_text
+    >>> rri = read_from_text('/path/to/file.txt')
+    >>> moving_median(rri)
+    RRi array([1114., 1113., 1113., 1066., ])
+    """
     return _moving_function(rri, order, np.median)
 
 
[…]
