### 2019-12-14 11e3699
Add docstring for detrend methods
