+++ hrv/detrend.py
@@ -6,7 +6,39 @@ from scipy.sparse import spdiags, dia_matrix
 from hrv.rri import RRiDetrended, RRi, _create_time_array
 
 
+__all__ = ['polynomial_detrend', 'smoothness_priors', 'sg_detrend']
+
+
 def polynomial_detrend(rri, degree=1):
+    """
+    Fits a polynomial function with degree=`degree` in the RRi series
+    and then subtract it from the signal.
+
+    Parameters
+    ----------
+    rri : array_like
+        sequence containing the RRi series
+    degree: integer, optional
+        degree of the polynomial function to be fitted in the RRi series.
+        Defaults to 1
+
+    Returns
+    -------
+    results : RRi array
+        instance of the RRi Detrended class containing the detrended RRi values
+
+    See Also
+    -------
+    smoothness_priors, sg_detrend
+
+    Examples
+    --------
+    >>> from hrv.detrend import polynomial_detrend
+    >>> from hrv.io import read_from_text
+    >>> rri = read_from_text('/path/to/file.txt')
+    >>> polynomial_detrend(rri)
+    RRi array([ 2.47769099e+01,  2.48650030e+01, -3.04917114e+02])
+    """
     if isinstance(rri, RRi):
         time = rri.time
         rri = rri.values
@@ -20,12 +52,61 @@ def polynomial_detrend(rri, degree=1):
 
 
 def smoothness_priors(rri, l=500, fs=4.0):
+    """
+    Estimates the stationary part of the resampled RRi series and subtracts it
+    from the signal. The RRi series must have equal time-space between
+    adjacents RRi values, therefore the original RRi is interpolated and then
+    resampled.
+
+    Parameters
+    ----------
+    rri : array_like
+        sequence containing the RRi series
+    l: integer, optional
+        the regularization parameter
+        Defaults to 500
+    fs: integer, optional
+        sampling frequency in each the RRi series will be resampled after
+        cubic interpolation. Defaults to 4
+
+    .. math::
+        The estimated stationary component of the RRi can represented as:
+        z_stat = (I - (l**2 @ D2.T @ D2)**2) @ z
+
+        where I is the identity matrix; D2 is the second order difference
+        matrix and z is the evenly spaced RRi series (containing both
+        stationaty and trend components)
+
+    Returns
+    -------
+    results : RRi array
+        instance of the RRi Detrended class containing the detrended RRi values
+
+    See Also
+    -------
+    polynomial_detrend, sg_detrend
+
+    References
+    ----------
+    - M.P.  Tarvainen,  P.O.  Ranta-aho  and  P.A.  Karjalainen,  An  advanced
+      detrending  method  with  application  to  HRV  analysis, IEEE
+      Transaction on Biomedical Engineering 49 (2002), 172–175.
+
+    Examples
+    --------
+    >>> from hrv.detrend import smoothness_priors
+    >>> from hrv.io import read_from_text
+    >>> rri = read_from_text('/path/to/file.txt')
+    >>> smoothness_priors(rri)
+    RRi array([  29.04627085,   42.17204395,   44.5743763])
+    """
     if isinstance(rri, RRi):
         time = rri.time
         rri = rri.values
     else:
         time = _create_time_array(rri)
 
+    # TODO: only interp if not interpolated yet
     cubic_spline = CubicSpline(time, rri)
     time_interp = np.arange(time[0], time[-1], 1.0/fs)
     rri_interp = cubic_spline(time_interp)
@@ -46,6 +127,42 @@ def smoothness_priors(rri, l=500, fs=4.0):
     )
 
 def sg_detrend(rri, window_length=51, polyorder=3,  *args, **kwargs):
+    """
+    Remove the low-frequency components of the RRi series with the low-pass
[…]
