### 2019-12-14 482ef70
Add dosctring for RRi classes
