+++ hrv/rri.py
@@ -1,3 +1,19 @@
+"""
+Objects for dealing with RR intervals series (RRi).
+
+This module offers two classes to represent RRi series as Python objects which
+brings widely used methods to extract information from tachograms.
+Once RRi or RRiDetrended are instanciated is it possible to calculate simple
+descriptive statistics of the RRi series, such: average, median, etc. It is
+also possible to visualize the series with the plot method, extract a smaller
+segment of the RRi series based on time information
+
+Classes
+------
+ - `RRi` -- RRi series class
+ - `RRiDetrended` -- detrended RRi values
+"""
+
 import sys
 
 from collections import MutableMapping, defaultdict
@@ -8,7 +24,35 @@ import numpy as np
 from .utils import _ellipsedraw
 
 
+__all__ = ['RRi', 'RRiDetrended']
+
+
 class RRi:
+    """An RRi series class.
+
+       The RRi class provides magic methods that make it instance behave like
+       a Python iterable. It also has methods that help to describe the
+       statistical properties of the instantiated RRi series and for
+       visualization purpose.
+
+       This class performs simple validation of the provided rri values:
+            - it can not have negative values
+            - if the median values are smaller than 10, it is presumed they
+              are in seconds and are transformed into miliseconds
+       If time information, which is optional, is provided it is also
+       validated:
+            - time array must have the same length as the RRi series
+            - must be monotonically increasing
+            - no negative values
+            - only the first value can be zero
+
+        When time information is not provided it will be created with the
+        cumulative sum of the RRi values:
+            time = np.cumsum(rri) / 1000.0
+            time -= time[0]  # to start at zero
+
+        Time is represented in seconds
+    """
     def __init__(self, rri, time=None, *args, **kwargs):
         if not isinstance(self, RRiDetrended):
             self.__rri = _validate_rri(rri)
@@ -36,25 +80,34 @@ class RRi:
 
     @property
     def values(self):
+        """Return a numpy array containing the RRi values"""
         return self.__rri
 
     @property
     def rri(self):
+        """Return a numpy array containing the RRi values"""
         return self.__rri
 
     @property
     def time(self):
+        """Return a numpy array containing the time information"""
         return self.__time
 
     @property
     def detrended(self):
+        """Return if the RRi series is detrended"""
         return self.__detrended
 
     @property
     def interpolated(self):
+        """Return if the RRi series is interpolated"""
         return self.__interpolated
 
     def describe(self):
+        """
+        Return a dictionary containing descriptive statistics from the RRi
+        series.
+        """
         table = _prepare_table(RRi(self.rri))
         rri_descr = RRiDescription(table)
         for row in table[1:]:
@@ -64,6 +117,10 @@ class RRi:
         return rri_descr
 
     def info(self):
+        """
+        Print information about RRi`s memory usage, length,
+        number of intervals, and if it is interpolated and/or detrended
+        """
         def _mem_usage(nbytes):
             mem_val = nbytes / 1024
             if mem_val < 1000:
@@ -91,19 +148,49 @@ class RRi:
         ))
 
     def to_hr(self):
+        """
+        Return a numpy array containing the heart rate calculated with
+        the RRi series
+        """
         return 60 / (self.rri / 1000.0)
 
     def time_range(self, start, end):
+        """
+        Crop the RRi series based in time information
+
+        Parameters
+        ----------
+        start : float
+            beginning of the new RRi series
+
+        end : float
+            end of the new RRi series
+        """
[…]
