+++ tox.ini
@@ -1,0 +1,12 @@
+[flake8]
+exclude =
+    # No need to traverse our git directory
+    .git,
+    # There's no value in checking cache directories
+    __pycache__,
+    __init__,
+    setup.py,
+    hrv/__init__.py,
+    docs/source/conf.py,
+    build,
+    dist,
