+++ .travis.yml
@@ -1,0 +1,12 @@
+language: python
+python:
+  - "3.7"
+install:
+  - pip install --upgrade pip
+  - pip install -r dev-requirements.txt
+script:
+  - make test
+
+branches:
+  only:
+    - feature/issue/docstrings
