### 2019-12-14 48d9e53
Add travis status image
