### 2019-12-17 6a6d851
Add codecov to project
Co-authored-by: samambaman <felipe.gomes.ferreira@gmail.com>
