+++ .travis.yml
@@ -6,3 +6,5 @@ install:
   - pip install -r dev-requirements.txt
 script:
   - make test
+after_success:
+    - codecov
