+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 python_files =  tests/test_*.py
+addopts = --cov=. 
