### 2019-12-18 7f2d01d
Add sample exercise rri
