### 2019-12-18 bd5f70d
Add sample exercise noisy rri
