+++ tests/test_load_sampledata.py
@@ -45,3 +45,16 @@ class TestLoadSampleData(TestCase):
             sample_rri[-3:],
             [704., 805., 808.]
         )
+
+    def test_load_sample_exercise_noisy_rri(self):
+        sample_rri = load_sample_data('exercise_noisy_rri.hrm')
+
+        self.assertIsInstance(sample_rri, RRi)
+        np.testing.assert_almost_equal(
+            sample_rri[:3],
+            [1971.,  985.,  995.]
+        )
+        np.testing.assert_almost_equal(
+            sample_rri[-3:],
+            [639., 642., 644.]
+        )
