### 2019-12-18 ae488bb
Create specific function to load rest rri
