### 2019-12-18 ca65140
Create specific function to load exercise rri
