+++ hrv/sampledata/__init__.py
@@ -1,0 +1,1 @@
+from hrv.sampledata._load import load_rest_rri, load_exercise_rri
