+++ tests/test_load_sampledata.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from hrv.io import RRi
 from hrv.sampledata._load import load_sample_data
-from hrv.sampledata import load_rest_rri
+from hrv.sampledata import load_rest_rri, load_exercise_rri
 
 
 class TestLoadSampleData(TestCase):
@@ -36,6 +36,19 @@ class TestLoadSampleData(TestCase):
 
     def test_load_sample_exercise_rri(self):
         sample_rri = load_sample_data('exercise_rri.hrm')
+
+        self.assertIsInstance(sample_rri, RRi)
+        np.testing.assert_almost_equal(
+            sample_rri[:3],
+            [1589.,  783.,  752.]
+        )
+        np.testing.assert_almost_equal(
+            sample_rri[-3:],
+            [562., 555., 557.]
+        )
+
+    def test_load_exercise_rri(self):
+        sample_rri = load_exercise_rri()
 
         self.assertIsInstance(sample_rri, RRi)
         np.testing.assert_almost_equal(
