### 2019-12-18 df8ef73
Create specific function to load noisy rri
