+++ hrv/sampledata/__init__.py
@@ -1,1 +1,5 @@
-from hrv.sampledata._load import load_rest_rri, load_exercise_rri
+from hrv.sampledata._load import (
+    load_rest_rri,
+    load_exercise_rri,
+    load_noisy_rri
+)
