### 2019-12-18 7ef344b
Remove noisy exercise rri
