+++ docs/source/firststeps.rst
@@ -1,0 +1,18 @@
+First Steps
+===========
+
+Installation
+############
+
+To install use pip:
+
+.. code-block:: bash
+
+   pip install hrv
+
+Or clone the repo:
+
+.. code-block:: bash
+
+   git clone https://github.com/rhenanbartels/hrv.git
+   python setup.py install
