### 2019-12-19 73b1e34
Add docstring for the sampledata functions
