### 2019-12-19 3f8c3eb
Use sampledata in docstrings
