### 2019-12-19 1b60fd4
Correct docs hyperlinks
