+++ docs/source/statistics.rst
@@ -63,3 +63,5 @@ RRi Basic Information
     Interpolated: False
     Detrended: False
     Memory Usage: 0.05Kb
+
+
