### 2019-12-19 d35a7ac
Add contribution section to docs
