### 2019-12-19 e8d2c68
Clean README and insert link to readthedocs
