### 2019-12-19 6317ef5
Add docs favicon
