+++ README.md
@@ -1,6 +1,10 @@
 [![Build Status](https://travis-ci.org/rhenanbartels/hrv.svg?branch=develop)](https://travis-ci.org/rhenanbartels/hrv)
 [![codecov](https://codecov.io/gh/rhenanbartels/hrv/branch/develop/graph/badge.svg)](https://codecov.io/gh/rhenanbartels/hrv)
-
+[![Current version at PyPI](https://img.shields.io/pypi/v/hrv.svg)](https://pypi.python.org/pypi/hrv)
+[![Downloads per month on PyPI](https://img.shields.io/pypi/dm/hrv.svg)](https://pypi.python.org/pypi/hrv)
+![Supported Python Versions](https://img.shields.io/pypi/pyversions/hrv.svg)
+![Software status](https://img.shields.io/pypi/status/rows.svg)
+[![License: LGPLv3](https://img.shields.io/pypi/l/hrv.svg)](https://github.com/rhenanbartels/hrv/blob/develop/LICENSE.md)
 
 
 Pythonic package for Heart Rate Variability Analysis
