### 2019-12-19 rhenanbartels/hrv#17 @rhenanbartels
Hi, @paulvangentcom,

Following your suggestions, I've added some features regarding the documentation:

 - A Docstring for every main function of the submodules
 - Add the project to **travis-ci** and add the badge to README
 - Add the project to **codecov** with **tests** coverage status and its respective badge
 - Add Badge telling the supported Python version
 - Moved the documentation to [readthedocs](https://hrv.readthedocs.io/en/latest/)

Regarding the paper, which improvements would you like to see in order to make it more close to the state of the art?

Thank you very much again for your suggestions. They improved a lot the quality of the hrv module;

Best Regards,

Rhenan
