### 2019-12-19 dbfd3a8
Set doc master page to conf.py
