+++ docs/source/conf.py
@@ -54,3 +54,4 @@ html_static_path = ['_static']
 
 html_logo = '../logo-small.png'
 html_favicon = '../favicon.png'
+master_doc = 'index'
