### 2019-12-24 rhenanbartels/hrv#17 @paulvangentcom
Hi @rhenanbartels,

Thanks for the work. The docs look nice and I think everything improved markedly!

For the paper, what I'm missing is:
- What other software is there and how is hrv different/complementary to what is there? A brief overview woudl suffice.
- The statement of need can be stronger: right now at the end of the summary section you describe what the software does, not why it's needed. I think the statement of need will flow naturally from the first point on overview of software.
- The nonlinear section is too brief. Describe what you're expressing (perhaps together with a plot): the standard deviation along and perpendicular to the identity line on an RR-plot. With poincare it's also customary to compute 's' (area of the ellipse described by sd1, sd2), and the sd1/sd2 ratio. As long as you link to the source, you can [nick the method here](https://github.com/paulvangentcom/heartrate_analysis_python/blob/master/heartpy/analysis.py#L637), and the [plotting method here](https://github.com/paulvangentcom/heartrate_analysis_python/blob/master/heartpy/visualizeutils.py#L189).

Good luck with the final touches!

-Paul
