### 2019-12-26 #1867 @Kevin-Mattheus-Moerman
@rhenanbartels can you respond to the reviewers comments/issues and summarize your actions here? Thanks :christmas_tree:
