### 2019-12-26 #1867 @rhenanbartels
Hi @Kevin-Mattheus-Moerman,  regarding [issue 16](https://github.com/rhenanbartels/hrv/issues/16) and  [Issue 17](https://github.com/rhenanbartels/hrv/issues/17),  I've implemented features for artifact and baseline trend removal, and I've worked on documentation-related and CI features, respectively:

## Already concluded

#### [Issue 16](https://github.com/rhenanbartels/hrv/issues/16) - Suggested by @hooman650
**Filter**
-  [threshold_filter](https://github.com/rhenanbartels/hrv/blob/ca28b2e6e08118e4ea4cb8d0e73b30a5c8313864/hrv/filters.py#L39): inspired in Kubios **threshold based artefact correction algorithm**, each RRi is compared to a local value consisting of the median of adjacent RRi. If the difference between RRi and the local median is greater than a given threshold in milliseconds this RRi is considered an ectopic beat. Ectopic RRi values are replaced with cubic spline interpolation of the entire tachogram.

**Trend removal**
  - [polynomial_detrend](https://github.com/rhenanbartels/hrv/blob/ca28b2e6e08118e4ea4cb8d0e73b30a5c8313864/hrv/detrend.py#L9): fits a polynomial model of a given degree to the tachogram and remove it from each RRi values.
  - [smoothness_priors](https://github.com/rhenanbartels/hrv/blob/ca28b2e6e08118e4ea4cb8d0e73b30a5c8313864/hrv/detrend.py#L22): widely used in hrv analyses, acts as a lowpass filter and remove complex trends from the RRi series
 - [sg_detrend](https://github.com/rhenanbartels/hrv/blob/ca28b2e6e08118e4ea4cb8d0e73b30a5c8313864/hrv/detrend.py#L48): uses the savitzky-golay lowpass filter to remove low-frequency trends from the RRi series

#### [Issue 17](https://github.com/rhenanbartels/hrv/issues/17) - Suggested by @paulvangentcom

 - A Docstring for every main function of the submodules
 - The project to **travis-ci** and added the badge to README
 - The project to **codecov** with **tests** coverage status and its respective badge
 - The Badge telling the supported Python version
 - Sample data: i.e from `hrv.sampledata import load_rest_rri`
 - Moved the documentation to [readthedocs](https://hrv.readthedocs.io/en/latest/)

## Next steps

During the following days, I will work on the improvements (suggested by @paulvangentcom) of the article's text and I will add the new features available in the 0.2.5 version of the `hrv` module, such:

 - [Add the Ponicaré Plot](https://hrv.readthedocs.io/en/latest/analysis.html#non-linear-analysis)
 - Add a Pre-processing section
 - Improve the Non-linear analysis section
 - Emphasize the importance of  the `hrv` module

Cheers,

Rhenan
