### 2019-12-29 59093bd
Add new sections and references
