+++ paper/paper.bib
@@ -167,4 +167,47 @@
   year={1990},
   publisher={Wiley Online Library}
 }
-
+@inproceedings{Gomes2019,
+   author = {Gomes, Pedro and Margaritoff, Petra and Silva, Hugo},
+   booktitle = {Proc. Int'l Conf. on Electrical, Electronic and Computing Engineering (IcETRAN)},
+   pages = {822-828},
+   title = {{pyHRV: Development and evaluation of an open-source python toolbox for heart rate variability (HRV)}},
+   year = {2019}
+}
+@inproceedings{rodriguez2008rhrv,
+  title={RHRV: An R-based software package for heart rate variability analysis of ECG recordings},
+  author={Rodr{\'\i}guez-Li{\~n}ares, Leandro and Vila, X and Mendez, A and Lado, M and Olivieri, D},
+  booktitle={3rd Iberian Conference in Systems and Information Technologies (CISTI 2008)},
+  pages={565--574},
+  year={2008}
+}
+@article{tarvainen2014kubios,
+  title={Kubios HRV--heart rate variability analysis software},
+  author={Tarvainen, Mika P and Niskanen, Juha-Pekka and Lipponen, Jukka A and Ranta-Aho, Perttu O and Karjalainen, Pasi A},
+  journal={Computer methods and programs in biomedicine},
+  volume={113},
+  number={1},
+  pages={210--220},
+  year={2014},
+  publisher={Elsevier}
+}
+@article{bartels2017sinuscor,
+  title={SinusCor: an advanced tool for heart rate variability analysis},
+  author={Bartels, Rhenan and Neumamm, Leonardo and Pe{\c{c}}anha, Tiago and Carvalho, Alysson Roncally Silva},
+  journal={Biomedical engineering online},
+  volume={16},
+  number={1},
+  pages={110},
+  year={2017},
+  publisher={BioMed Central}
+}
+@article{tarvainen2002advanced,
+  title={An advanced detrending method with application to HRV analysis},
+  author={Tarvainen, Mika P and Ranta-Aho, Perttu O and Karjalainen, Pasi A},
+  journal={IEEE Transactions on Biomedical Engineering},
+  volume={49},
+  number={2},
+  pages={172--175},
+  year={2002},
+  publisher={IEEE}
+}
