+++ paper/paper.md
@@ -45,16 +45,24 @@ Roughly, frequency domain analysis involves the calculation of the spectral ener
 | LF<sub>n.u</sub> | normalized units $\frac{LF}{Total Power - VLF}$ | |
 | HF<sub>n.u</sub> | normalized units $\frac{HF}{Total Power - VLF}$ | |
 
-Non-linear indices are also frequently used to extract information from the ANS based on the heart rate fluctuations patterns. SD1 and SD2 respectively reflect the short and long term fluctuations of heart rate, which can be calculated from the Poincaré ellipse plot [@berntson1997heart] or derived from SDSD and SDNN values as shown by equations 2 and 3 below. 
+Non-linear indices are also frequently used to extract information from the ANS based on the heart rate fluctuations patterns. SD1 reflects the short term fluctuations of the heart rate, and for this reason is highly correlated with the RMSSD, pNN50 and HF indexes. The SD2 index reflects both short and long terms of the fluctuation of the heart rate and correlates with SDNN and LF indexes. Both SD1 and SD2 are generally represented using the Poincaré ellipse plot [@berntson1997heart]. In this plot, the SD1 index represents the standard deviation spread orthogonally to the identity line (y=x) and it is the ellipse width, while the SD2 index represents the standard deviation spread along the identity line and specifies the length of the ellipse. At the end of the following section, the Poincaré plot of a given RRi series is depicted using the module presented in the current article. 
+
+The calculation of the SD1 and SD2 indexes can be derived from SDSD and SDNN values as shown by equations 2 and 3 below:
+
 $$SD1 = \sqrt{2SDNN^{2} - 2SD2^{2}}$$ - Equation 2
 
 $$SD2 = \sqrt{2SDNN^{2} - \frac{1}{2}SDSD^{2}}$$ - Equation 3
 
-The hrv is a simple and open-source Python module that brings the most widely used techniques to work with the RRi series and HRV analyses without losing the power and flexibility of a native Python object and a numpy array [@numpy]. The following sections present the basic workflow with RRi series, starting with reading a file containing a tachogram, visualizing the given RRi series, dealing with noise filtering and, finally, calculating time/frequency domain and non-linear HRV indices.
+There are several software packages written in many different programming languages that offer functions to work with RRi signals. Some of them have a command-line-based interface [@rodriguez2008rhrv], and others offer a user's interface with and all types of widgets to improve the interaction with the RRi series and with the analyses  [@tarvainen2014kubios,@bartels2017sinuscor]. 
+Specifically for Python, there is also a myriad of open-source packages available and ready to be used to work with HRV analisys, such as [hrvanalysis](https://github.com/Aura-healthcare/hrvanalysis), pyhrv [@Gomes2019], and [heartpy](https://github.com/paulvangentcom/heartrate_analysis_python). Although these modules do a great work offering many of the most widely used techniques to deal with and to extract relevant information from HRV signals, their functions interface (API) relies on RRi signals stored as Python iterable or numpy arrays and is based mostly on the procedural programming paradigm. 
+
[…]
