### 2020-01-04 b32ae62
Improve text and add poincare and detrended rri figures
