+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Heart rate variability (HRV) is a non-invasive tool to assess the cardiac autonomic integrity and cardiovascular homeostasis [@electrophysiology1996heart]. HRV quantifies the instantaneous variations in the RR intervals, which is produced by the balanced action of the parasympathetic and sympathetic branches of the autonomic nervous system (ANS) over the sinoatrial (SA) node, and modulated by different physiological inputs (e.g., respiration, blood pressure, temperature, emotions, etc) [@malik1990heart]. Specifically, the parasympathetic activation produces fast and short-lasting bradycardia, which results in high-frequency oscillations (i.e., 0.15 - 0.40 Hz) in heart rate. On the other hand, sympathetic activation produces slower and longer-lasting variations in heart rate, which results in low-frequency (i.e., ~ 0.1 Hz) oscillations in heart rate.  
+Heart rate variability (HRV) is a non-invasive tool to assess the cardiac autonomic integrity and cardiovascular homeostasis [@electrophysiology1996heart]. HRV quantifies the instantaneous variations in the RR intervals (RRi), which is produced by the balanced action of the parasympathetic and sympathetic branches of the autonomic nervous system (ANS) over the sinoatrial (SA) node, and modulated by different physiological inputs (e.g., respiration, blood pressure, temperature, emotions, etc) [@malik1990heart]. Specifically, the parasympathetic activation produces fast and short-lasting bradycardia, which results in high-frequency oscillations (i.e., 0.15 - 0.40 Hz) in heart rate. On the other hand, sympathetic activation produces slower and longer-lasting variations in heart rate, which results in low-frequency (i.e., ~ 0.1 Hz) oscillations in heart rate.  
 
 Increased HRV indicates a predominance of the parasympathetic over the sympathetic activation in the SA node, and indicates enhanced cardiac autonomic flexibility and improved overall health [@malik1990heart]. Conversely, a reduced HRV is usually accompanied by sympathetic dominance and suggests increased rigidity and loss of control of the ANS to the cardiovascular system, which is a sign of disease vulnerability [@malik1990heart]. For instance, a reduced SDNN (i.e., standard deviation of RR intervals, a simple statistical-based time domain  index of HRV) has been shown to overperform traditional cardiovascular risk parameters in predicting mortality in a cohort of heart failure patients [@nolan1998prospective]. A reduced HRV has also been linked with metabolic dysfunction [@weissman2006power], increased inflammation [@sajadieh2004increased], depression [@sgoifo2015autonomic],  psychiatric disorders [@degiorgio2010rmssd], sleep disturbance [@burton2010reduced], among others. Based on this wide prognostic utility, the interest in approaches to evaluate HRV has shown an exponential growth in different medicine specialties and research fields in the recent years.  
 
@@ -45,20 +45,22 @@ Roughly, frequency domain analysis involves the calculation of the spectral ener
 | LF<sub>n.u</sub> | normalized units $\frac{LF}{Total Power - VLF}$ | |
 | HF<sub>n.u</sub> | normalized units $\frac{HF}{Total Power - VLF}$ | |
 
[…]
