### 2020-01-04 c94cf7f
Final touches in article's text
