### 2020-01-04 #1867 @rhenanbartels
Hi @Kevin-Mattheus-Moerman,

    I've just put the finishing touches on the article`s text. It is already available in the develop branch in the repo.
    What is the next step?

Best,
Rhenan
