### 2020-01-11 #1867 @Kevin-Mattheus-Moerman
@hooman650, @PGomes92, @paulvangentcom thanks for your help so far. The author has made various changes to address your comments. Can you please pick up the review at this point and review these changes? Please summarize any remaining issues you might have. Thanks!
