### 2020-02-05 rhenanbartels/hrv#17 @paulvangentcom
Thanks for all the updates Rhenan. I think it improved a lot.

Two final comments:
- The community guidelines on the docs on how to contribute now only outline the creation of a virtualenv. Including something about things like the preferred coding style, docstring style, and module structure might help people starting out with contributing.
- You link to the repo for HeartPy. We prefer you link to [our software paper instead](https://openresearchsoftware.metajnl.com/articles/10.5334/jors.241/) as it contains more background info.

Cheers,
Paul
