### 2020-02-05 #1867 @paulvangentcom
@Kevin-Mattheus-Moerman yes these are on the repo, appended to the [same issue as before](https://github.com/rhenanbartels/hrv/issues/17)
