### 2020-02-12 54f2f27
Start adding coding style do contribution docs
