+++ paper/paper.md
@@ -54,7 +54,7 @@ $$SD1 = \sqrt{2SDNN^{2} - 2SD2^{2}}$$ - Equation 2
 $$SD2 = \sqrt{2SDNN^{2} - \frac{1}{2}SDSD^{2}}$$ - Equation 3
 
 There are several software packages written in many different programming languages that offer functions to work with RRi signals. Some of them have a command-line interface [@rodriguez2008rhrv] and others offer a user's interface to improve the interaction with the RRi series and the analyses  [@tarvainen2014kubios,@bartels2017sinuscor]. 
-Specifically for Python, there is also open-source packages available and ready to work on HRV analisys, such as [hrvanalysis](https://github.com/Aura-healthcare/hrvanalysis), pyhrv [@Gomes2019], and [heartpy](https://github.com/paulvangentcom/heartrate_analysis_python). Although these modules do a great work offering many of the most widely used techniques to deal with tachograms and to extract relevant information from HRV signals, their functions interface (API) relies on RRi signals stored as Python iterable or numpy arrays and is based mostly on the procedural programming paradigm. 
+Specifically for Python, there is also open-source packages available and ready to work on HRV analisys, such as [hrvanalysis](https://github.com/Aura-healthcare/hrvanalysis), pyhrv [@Gomes2019], and heartpy [@van2019analysing]. Although these modules do a great work offering many of the most widely used techniques to deal with tachograms and to extract relevant information from HRV signals, their functions interface (API) relies on RRi signals stored as Python iterable or numpy arrays and is based mostly on the procedural programming paradigm. 
 
 The `hrv` is a simple and open-source Python module that comes with the most common techniques for filtering, detrending and extracting information about the ANS from the RRi signals without losing the power and flexibility of a native Python object and a numpy arrays [@numpy]. It brings the necessary methods to work with a tachogram encapsulated in a Python class. In other words, once an RRi class is instantiated there are several methods available for visualization, descriptive statistics, slicing the signal in shorter segments, and displaying the metadata of the series.
 
