+++ paper/paper.bib
@@ -211,3 +211,12 @@
   year={2002},
   publisher={IEEE}
 }
+@article{van2019analysing,
+  title={Analysing noisy driver physiology real-time using off-the-shelf sensors: heart rate analysis software from the taking the fast lane project},
+  author={van Gent, Paul and Farah, Haneen and van Nes, Nicole and van Arem, Bart},
+  journal={Journal of Open Research Software},
+  volume={7},
+  number={1},
+  year={2019},
+  publisher={Ubiquity Press}
+}
