### 2020-02-20 71349a9
Add docstring example in how to contribute
