### 2020-02-20 rhenanbartels/hrv#17 @rhenanbartels
Hi @paulvangentcom, I've just implemented the suggestion you've made.

 - I've Created a section telling what is the preferred docstring styles. It is still simple, but it will improve as the project evolves.

 - I've Changed the citation's link to HeartPy. By the way, it is a very nice article.

Thank you very much!

Cheers,

Rhenan
