### 2020-02-20 #1867 @rhenanbartels
Hi @Kevin-Mattheus-Moerman, I've just pushed the suggestions made by @paulvangentcom.

Cheers,

Rhenan
