### 2020-03-06 #1867 @paulvangentcom
Yes agreed, the changes are good as far as I'm concerned. Thanks for guiding the process @Kevin-Mattheus-Moerman , and thanks for the work @rhenanbartels
