### 2020-06-10 9c3cfcb
Start method to split rri in time
