+++ hrv/sampledata/_load.py
@@ -19,7 +19,7 @@ def load_sample_data(filename):
         rri = load_sample_data('rest_rri.txt')
     """
     extension = os.path.splitext(filename)[1]
-    handler = {'.txt': read_from_text, '.hrm': read_from_hrm}
+    handler = {".txt": read_from_text, ".hrm": read_from_hrm}
 
     here = os.path.dirname(__file__)
     complete_path = os.path.join(here, filename)
@@ -39,7 +39,7 @@ def load_rest_rri():
     >>> load_rest_rri()
     RRi array([1114., 1113., ..., 1066., 1119.])
     """
-    return load_sample_data('rest_rri.txt')
+    return load_sample_data("rest_rri.txt")
 
 
 def load_exercise_rri():
@@ -61,7 +61,7 @@ def load_exercise_rri():
     >>> load_exercise_rri()
     RRi array([1589.,  783.,  752., ...,  562.,  555.,  557.])
     """
-    return load_sample_data('exercise_rri.hrm')
+    return load_sample_data("exercise_rri.hrm")
 
 
 def load_noisy_rri():
@@ -84,4 +84,4 @@ def load_noisy_rri():
     >>> load_noisy_rri()
     RRi array([904., 913., 937., ..., 704., 805., 808.])
     """
-    return load_sample_data('noisy_rri.hrm')
+    return load_sample_data("noisy_rri.hrm")
