+++ hrv/sampledata/__init__.py
@@ -1,5 +1,1 @@
-from hrv.sampledata._load import (
-    load_rest_rri,
-    load_exercise_rri,
-    load_noisy_rri
-)
+from hrv.sampledata._load import load_rest_rri, load_exercise_rri, load_noisy_rri

[…]
