### 2020-06-10 54f648f
Correct assert clause
