### 2020-06-11 89599a2
Keep last segment even if it is shorter
