### 2020-06-11 bd68e3e
Start module to calculate nonstationary indices
