### 2020-06-11 cea3072
Return time-varying indexes
