+++ hrv/nonstationary.py
@@ -20,6 +20,12 @@ class TimeVarying:
             key: [item[key] for item in results] for key in results[0].keys()
         }
 
+    def __getattr__(self, index):
+        try:
+            return self.transponsed[index]
+        except KeyError:
+            raise ValueError(f"index `{index}` does not exist.")
+
     def plot(self, index="rmssd"):
         fig, ax = plt.subplots(1, 1)
         ax.plot(self.transponsed[index])
