### 2020-06-12 8896f80
Plot time-varying indexes
