### 2020-06-17 930d3a3
Improve time varying plot
