### 2020-06-17 6b70903
Define TV __str__ method
