### 2020-06-20 4c63bfa
Plot TimeVarying object
