+++ hrv/nonstationary.py
@@ -1,6 +1,7 @@
 from collections import defaultdict
 
 import matplotlib.pyplot as plt
+import numpy as np
 
 from hrv.classical import time_domain
 from hrv.rri import RRi
@@ -10,17 +11,16 @@ __all__ = ["time_varying"]
 
 
 class TimeVarying:
-    def __init__(self, results, rri_segments, seg_size, overlap):
+    def __init__(self, rri, results, rri_segments, seg_size, overlap):
+        self.rri = rri
         self.results = results
         self.transponsed = self._transform_results(results)
         self.rri_segments = rri_segments
-        self.seg_size=seg_size
-        self.overlap=overlap
+        self.seg_size = seg_size
+        self.overlap = overlap
 
     def _transform_results(self, results):
-        return {
-            key: [item[key] for item in results] for key in results[0].keys()
-        }
+        return {key: [item[key] for item in results] for key in results[0].keys()}
 
     def ylabel_mapper(self, index):
         mapper = {
@@ -34,6 +34,9 @@ class TimeVarying:
         }
         return mapper.get(index)
 
+    def build_xaxis(self):
+        return [np.median(r.time) for r in self.rri_segments]
+
     def __getattr__(self, index):
         try:
             return self.transponsed[index]
@@ -45,12 +48,12 @@ class TimeVarying:
         if ax is None:
             fig, ax = plt.subplots(1, 1)
 
-        ax.plot(self.__getattr__(index), *args, **kwargs)
+        ax.plot(self.build_xaxis(), self.__getattr__(index), *args, **kwargs)
         ax.set(xlabel="Time Interval (s)", ylabel=self.ylabel_mapper(index))
         plt.show(block=False)
 
         return fig, ax
-    
+
     def __str__(self):
         return f"Time Varying {self.seg_size}:{self.overlap} - #{len(self.results)}"
 
@@ -59,13 +62,9 @@ def time_varying(rri, seg_size, overlap, keep_last=False):
     if not isinstance(rri, RRi):
         rri = RRi(rri)
 
-    segments = rri.time_split(
-        seg_size=seg_size,
-        overlap=overlap,
-        keep_last=keep_last
-    )
+    segments = rri.time_split(seg_size=seg_size, overlap=overlap, keep_last=keep_last)
     results = []
     for segment in segments:
         results.append(time_domain(segment))
 
-    return TimeVarying(results, segments, seg_size=seg_size, overlap=overlap)
+    return TimeVarying(rri, results, segments, seg_size=seg_size, overlap=overlap)
