### 2020-06-20 935eaf1
Plot TV with RRi
