### 2020-06-27 68f7928
Improve time varying docs
