+++ docs/source/nonstationary.rst
@@ -10,13 +10,21 @@ Without running analysis and only visually inspecting the time series, is possib
 to tell that the average and the standard deviation of the RRi are not constant
 as a function of time.
 
-
-In order to extract useful information about the dynamics of non-stationary RRi series,
-the following methods applies the classical metrics in shorter running adjacent segments.
-
 .. image:: ../figures/exercise_rri.png
     :width: 500 px
 
+In order to extract useful information about the dynamics of non-stationary RRi series,
+the following methods applies the classical metrics in shorter running adjacent segments, 
+as illustrated in the following image:
+
+.. image:: ../figures/sliding_segments.png
+    :width: 500 px
+
+For example, for a segment size of **30s** (S) and **15s** (O) overlap a signal with **300s** (D) will have P segments:
+
+P = int((D - S) / (S - O)) + 1
+
+P = int((300 - 30) / (30 - 15)) + 1 = 19 segments
 
 Time Varying
 ############
@@ -56,17 +64,3 @@ Short Time Fourier Transform
 ############################
 
 To be implemented.
-
-
-
-.. image:: ../figures/rri_range.png
-    :width: 500 px
-
-Time offset can be reset from the RRi series range:
-
-.. code-block:: python
-
-    rri_range.reset_time(inplace=True)
-
-.. image:: ../figures/rri_range_reset.png
-    :width: 500 px
