### 2020-07-08 #1867 @Kevin-Mattheus-Moerman
@rhenanbartels I added myself as reviewer since the other reviewers dropped out. I installed the software successfully and have started testing. A couple of boxes are unticked.

**Documentation**
- [ ] **A statement of need**_: Do the authors clearly state what problems the software is designed to solve and who the target audience is?_
 Both the readme and the first page of the documentation simply state:
`hrv is a simple Python module that brings the most widely used techniques to work with RRi series and Heart Rate Variability (HRV) analyses without losing the Power and Flexibility of a native Python object and numpy arrays.`
Can you expand both to more clearly state what the software functionality addresses?

- [ ] Although the documentation seems to cover all functionality it seems to contain mainly snippets of code. Perhaps I am not interacting with your documentation properly but to test things I found myself copy-pasting these into my Spyder IDE to run them. Could you offer test scripts to users in a more direct way? It would be great to have files that test the functionality fully which users can run. Also some of the documentation contains this as example data:
```python
rri = RRi([800, 810, 815, 750, 753, 905])
```
This works for showing how to use the code components but is quite limited for testing the functionality. Could you add in the tests some more advanced examples containing larger test data sets? Or define it more clearly if you have it already?
This is just a suggestion but would it be easy to create a single Jupyter notebook that covers all functionality (perhaps combining the code snippets from the documentation) which also loads example datasets?

- [ ] Can you tell me where to find and study automated tests? Perhaps consider my above comments and add a `testing` section to the documentation. I see there is a "running the tests" section, but this seems to be for more advanced users. Perhaps consider (recommendation not requirement) add (e.g. in the form of a jupyter script) a simple test set and demo for new users?

Thanks
