### 2020-07-08 #1867 @Kevin-Mattheus-Moerman
@rhenanbartels apologies for the delay with this submission. With the pandemic our journal, editors (myself included), and reviewers have found it challenging to keep things going at a normal pace. Your submission was particularly "badly hit" it seems as unfortunately two reviewers became unresponsive.
I just wanted to apologize for this and add that I am committed to provide a fast review here and to also stay on top of editing this submission as fast as possible.
