### 2020-07-12 #1867 @rhenanbartels
Hi @Kevin-Mattheus-Moerman, yeah!

Thank you very much for the comments. I really liked the idea of creating a Jupyter Notebook with some real use cases. I am going to work on it and on the other suggestions during the following week.

Cheers!
