### 2020-07-16 f2da530
Fix rri time duration info
