### 2020-07-16 #1867 @rhenanbartels
Hi, @Kevin-Mattheus-Moerman

I've just added a [Jupyter notebook](https://github.com/rhenanbartels/hrv/tree/develop/notebooks) containing a walkthrough with the main steps of the pipeline for analyzing RRi/HRV signals (the same first steps we use in our studies in our Lab):

- Reading a file from the File System containing the information of the fluctuations of the heartbeats;
- Inspecting and visualizing the signal to check if everything is according to the expected;
- Filtering ectopic beats;
- Selecting two smaller segments of the filtered signal;
- Extracting Time and Frequency Domain features;
- Comparing the results extracted from each segment.

Additionally, there is also an example of one of the techniques applied to extract information about the dynamics of non-stationary RRi signals.

I will create more Jupyter notebooks as new features are added to the module. It is a great way to show its application in real scenarios and to be more intuitive for new users.

#

Regarding the `statement of need`, I've added, on both Docs and README, a complementary sentence that expands the description of the software functionality, which is: a module for inspecting, visualizing, pre-process, and analyze HRV analyses written in idiomatic Python.

Thank you very much for your suggestion and let me know if any other update is needed.

Cheers,

Rhenan
