### 2020-07-22 #1867 @Kevin-Mattheus-Moerman
@rhenanbartels thanks. I checked out the Jupyter notebook. It looks great! I do get an error on the last two cells:
![Screenshot from 2020-07-22 08-50-30](https://user-images.githubusercontent.com/8392709/88149882-94fa9700-cbf8-11ea-89a1-dcc8e8ebcfad.png)

Any clue as to what is going on?
