### 2020-07-22 #1867 @Kevin-Mattheus-Moerman
@rhenanbartels can you check/add those missing DOI's :point_up:
