### 2020-07-22 b610822
Add DOI's to paper's references
