+++ paper/paper.bib
@@ -7,7 +7,8 @@
   number={15},
   pages={1510--1516},
   year={1998},
-  publisher={Am Heart Assoc}
+  publisher={Am Heart Assoc},
+  doi={10.1161/01.cir.98.15.1510}
 }
 @article{weissman2006power,
   title={Power spectral analysis of heart rate variability during the 100-g oral glucose tolerance test in pregnant women},
@@ -17,7 +18,8 @@
   number={3},
   pages={571--574},
   year={2006},
-  publisher={Am Diabetes Assoc}
+  publisher={Am Diabetes Assoc},
+  doi={https://doi.org/10.2337/diacare.29.03.06.dc05-2009}
 }
 @article{sajadieh2004increased,
   title={Increased heart rate and reduced heart-rate variability are associated with subclinical inflammation in middle-aged and elderly subjects with no apparent heart disease},
@@ -27,7 +29,8 @@
   number={5},
   pages={363--370},
   year={2004},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1016/j.ehj.2003.12.003}
 }
 @article{sgoifo2015autonomic,
   title={Autonomic dysfunction and heart rate variability in depression},
@@ -37,7 +40,8 @@
   number={3},
   pages={343--352},
   year={2015},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.3109/10253890.2015.1045868}
 }
 @article{degiorgio2010rmssd,
   title={RMSSD, a measure of vagus-mediated heart rate variability, is associated with risk factors for SUDEP: the SUDEP-7 Inventory},
@@ -47,7 +51,8 @@
   number={1},
   pages={78--81},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.yebeh.2010.06.011}
 }
 @article{burton2010reduced,
   title={Reduced heart rate variability predicts poor sleep quality in a case--control study of chronic fatigue syndrome},
@@ -57,7 +62,8 @@
   number={1},
   pages={71--78},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00221-010-2296-1}
 }
 @article{welch1967use,
   title={The use of fast Fourier transform for the estimation of power spectra: a method based on time averaging over short, modified periodograms},
@@ -67,7 +73,8 @@
   number={2},
   pages={70--73},
   year={1967},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TAU.1967.1161901}
 }
 
 @article{electrophysiology1996heart,
@@ -78,7 +85,8 @@
   number={5},
   pages={1043--1065},
   year={1996},
-  publisher={Am Heart Assoc}
+  publisher={Am Heart Assoc},
+  doi={https://doi.org/10.1161/01.CIR.93.5.1043}
 }
 @Misc{numpy,
   author =    {Travis Oliphant},
@@ -96,7 +104,8 @@
   volume={11},
   number={1},
   pages={39--48},
-  year={2005}
+  year={2005},
+  doi={10.12921/cmst.2005.11.01.39-48}
 }
 @article{berntson1997heart,
   title={Heart rate variability: origins, methods, and interpretive caveats},
@@ -106,16 +115,8 @@
   number={6},
   pages={623--648},
   year={1997},
-  publisher={Wiley Online Library}
-}
-@article{piskorski2005filtering,
-  title={Filtering poincare plots},
-  author={Piskorski, Jaros{\l}aw and Guzik, Przemys{\l}aw},
[…]
