### 2020-07-22 #1867 @rhenanbartels
Hi @Kevin-Mattheus-Moerman, the `hrv.nonstationary` is a WIP added in version `0.2.8`. I guess `pip install -U hrv` would do the trick.

Regarding the missing DOI`s, I've added them in the last commit.
Let me know if there is anything else missing.

Cheers
