### 2020-07-23 #1867 @Kevin-Mattheus-Moerman
> pip install -U hrv

Yep that did the trick.
