### 2020-07-23 #1867 @Kevin-Mattheus-Moerman
@rhenanbartels once the paper compiles (you can run `@whedon generate pdf` to check compilation) we are ready to process this paper for acceptance in JOSS. Can you do the following:

- [ ] Check those invalid DOI's :point_up:  you may need to remove the `https://doi.org/` prefix
- [ ] Since your paper is about to be processed for acceptance, I suggest you proofread it once more and also that you verify that the author names, affiliations, and the acknowledgements are accurate/complete.
- [ ] Can you archive a copy of the reviewed software on ZENODO and report back here with the archive DOI? (some find this automated approach useful https://guides.github.com/activities/citable-code/). For the archive please make sure the meta-data matches that of the paper e.g. in terms of author names and the title (you can amend this if needed).
- [ ] Confirm the version of the reviewed and archived software? Is it still at `0.2.4`?
