### 2020-07-24 349fbd2
Correct some DOIs
