+++ paper/paper.bib
@@ -19,7 +19,7 @@
   pages={571--574},
   year={2006},
   publisher={Am Diabetes Assoc},
-  doi={https://doi.org/10.2337/diacare.29.03.06.dc05-2009}
+  doi={10.2337/diacare.29.03.06.dc05-2009}
 }
 @article{sajadieh2004increased,
   title={Increased heart rate and reduced heart-rate variability are associated with subclinical inflammation in middle-aged and elderly subjects with no apparent heart disease},
@@ -86,7 +86,7 @@
   pages={1043--1065},
   year={1996},
   publisher={Am Heart Assoc},
-  doi={https://doi.org/10.1161/01.CIR.93.5.1043}
+  doi={10.1161/01.CIR.93.5.1043}
 }
 @Misc{numpy,
   author =    {Travis Oliphant},
@@ -168,7 +168,7 @@
   pages={570--576},
   year={1990},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/clc.4960130811}
+  doi={10.1002/clc.4960130811}
 }
 @inproceedings{Gomes2019,
    author = {Gomes, Pedro and Margaritoff, Petra and Silva, Hugo},
@@ -193,7 +193,7 @@
   pages={210--220},
   year={2014},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.cmpb.2013.07.024}
+  doi={10.1016/j.cmpb.2013.07.024}
 }
 @article{bartels2017sinuscor,
   title={SinusCor: an advanced tool for heart rate variability analysis},
@@ -204,7 +204,7 @@
   pages={110},
   year={2017},
   publisher={BioMed Central},
-  doi={https://doi.org/10.1186/s12938-017-0401-4}
+  doi={10.1186/s12938-017-0401-4}
 }
 @article{tarvainen2002advanced,
   title={An advanced detrending method with application to HRV analysis},
@@ -225,5 +225,5 @@
   number={1},
   year={2019},
   publisher={Ubiquity Press},
-  doi={http://doi.org/10.5334/jors.241}
+  doi={10.5334/jors.241}
 }
