### 2020-07-25 #1867 @rhenanbartels
@Kevin-Mattheus-Moerman
>     * [x]  Check those invalid DOI's ☝️  you may need to remove the `https://doi.org/` prefix
>
>     * [x]  Since your paper is about to be processed for acceptance, I suggest you proofread it once more and also that you verify that the author names, affiliations, and the acknowledgements are accurate/complete.
>
>     * [X]  Can you archive a copy of the reviewed software on ZENODO and report back here with the archive DOI? (some find this automated approach useful https://guides.github.com/activities/citable-code/). For the archive please make sure the meta-data matches that of the paper e.g. in terms of author names and the title (you can amend this if needed).
>
>     * [ ]  Confirm the version of the reviewed and archived software? Is it still at `0.2.4`?

The Zenodo's link https://zenodo.org/record/3960216

Regarding the module version, we started at version 0.2.4, but throughout the review process the `hrv` package was updated to version 0.2.8, which is the latest release (archived in Zenodo).
