### 2020-07-25 #1867 @rhenanbartels
@Kevin-Mattheus-Moerman I am not experienced with Zenodo, so what will happen if we publish a new version of the software, let's say version `0.2.9`. Will the [link](https://zenodo.org/record/3960216) break?
