### 2020-07-25 #1867 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman I am not experienced with Zenodo, so what will happen if we publish a new version of the software, let's say version `0.2.9`. Will the [link](https://zenodo.org/record/3960216) break?

If you chose this approach https://guides.github.com/activities/citable-code/ and turned on the automatic system you should get a separate and new DOI for each release. So the new releases will not disturb this archived version.
